ujBilingualLab/requirements.txt
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-speech==0.36.3
googleapis-common-protos==1.5.8
pandas==0.23.4
SpeechRecognition==3.8.1